+++ packages/backend/src/nest/connections-manager/connections-manager.service.ts
@@ -51,7 +51,7 @@ import {
   InvitationData,
   SetUserProfileResponse,
 } from '@quiet/types'
-import { CONFIG_OPTIONS, QSS_ENABLED, QSS_ENDPOINT, SERVER_IO_PROVIDER, SOCKS_PROXY_AGENT } from '../const'
+import { CONFIG_OPTIONS, QSS_ALLOWED, QSS_ENDPOINT, SERVER_IO_PROVIDER, SOCKS_PROXY_AGENT } from '../const'
 import { Libp2pService } from '../libp2p/libp2p.service'
 import { CreatedLibp2pPeerId, Libp2pEvents, Libp2pNodeParams, Libp2pPeerInfo } from '../libp2p/libp2p.types'
 import { LocalDbService } from '../local-db/local-db.service'
@@ -88,7 +88,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     @Inject(SERVER_IO_PROVIDER) public readonly serverIoProvider: ServerIoProviderTypes,
     @Inject(CONFIG_OPTIONS) public configOptions: ConfigOptions,
     @Inject(SOCKS_PROXY_AGENT) public readonly socksProxyAgent: Agent,
-    @Inject(QSS_ENABLED) private readonly qssEnabled: boolean,
+    @Inject(QSS_ALLOWED) private readonly qssAllowed: boolean,
     @Inject(QSS_ENDPOINT) private readonly qssEndpoint: string | undefined,
     private readonly socketService: SocketService,
     public readonly libp2pService: Libp2pService,
@@ -218,7 +218,10 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       try {
         this.logger.info('Loading sigchain for community', community.name)
         const loadedSigchain = await this.sigChainService.loadChain(community.name, true)
-        const connected = await this.qssService.connect(!!community.qssEnabled, community.qssEndpoint)
+        if (community.qssEnabled) {
+          this.qssService.enableForCommunity(loadedSigchain.team!.id)
+        }
+        const connected = await this.qssService.connect(community.qssEndpoint)
         if (connected) {
           await this.qssService.signInToCommunity(loadedSigchain.team!.id, loadedSigchain)
         }
@@ -370,10 +373,10 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     }
   }
 
-  private async createCommunityOnQss(community: Community, sigchain: SigChain): Promise<void> {
-    const connected = await this.qssService.connect(this.qssEnabled, this.qssEndpoint)
+  private async createCommunityOnQss(sigchain: SigChain): Promise<void> {
+    const connected = await this.qssService.connect(this.qssEndpoint)
     if (connected) {
-      await this.qssService.createCommunity(community, sigchain)
+      await this.qssService.createCommunity(sigchain)
     }
   }
 
@@ -404,7 +407,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       psk: Libp2pService.generateLibp2pPSK().psk,
       ownership: CommunityOwnership.Owner,
       teamId: sigchain.team!.id,
-      qssEnabled: this.qssEnabled,
+      qssEnabled: this.qssAllowed,
       qssEndpoint: this.qssEndpoint,
     }
 
@@ -412,7 +415,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     await this.localDbService.setCurrentCommunityId(community.id)
 
     // purposely don't await
-    this.createCommunityOnQss(community, sigchain)
+    this.createCommunityOnQss(sigchain)
 
     await this.launchCommunity(community)
 
@@ -441,7 +444,8 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       inviteData.authData.teamId != null &&
       inviteData.qssEndpoint != null
     ) {
-      const connected = await this.qssService.connect(true, inviteData.qssEndpoint)
+      this.qssService.enableForCommunity(inviteData.authData.teamId)
+      const connected = await this.qssService.connect(inviteData.qssEndpoint)
       if (connected) {
         await this.qssService.signInToCommunity(inviteData.authData.teamId, sigChain, inviteData.authData.communityName)
       }